# Optimize EncryptParameterRewriterBuilder and EncryptPredicateColumnTokenGenerator logic